更新

+++ SteamSSFN/Form1.cs
@@ -48,18 +48,23 @@ namespace SteamSSFN
             return (string)key.GetValue("SteamPath"); ;
         }
 
-        private void button1_Click(object sender, EventArgs e){
-            if (File.Exists("temp")) {
-                File.Delete("temp");
-            }
+        //写入授权
+        private void button1_Click(object sender, EventArgs e)
+        {
+            if (File.Exists("temp")) File.Delete("temp");
+            //获取ssfn下载地址
             HttpDownloadFile("https://ssfnbox.com/download/" + textBox1.Text, "temp");
             string ssfnURL = GetBetweenStr(File.ReadAllLines("temp")[12], "window.location.assign(\"", "\"");
-            if (string.IsNullOrEmpty(ssfnURL)) {
+            if (string.IsNullOrEmpty(ssfnURL))
+            {
                 MessageBox.Show("未找到SSFN", "错误");
             }
-            else {
+            else
+            {
+                //判断steam目录是否正确
                 if (File.Exists(textBox2.Text + "/steam.exe"))
                 {
+                    //在steam根目录下搜索并删除ssfn
                     string[] searchfile = Directory.GetFiles(textBox2.Text, "ssfn*").Select(path => Path.GetFileName(path)).ToArray();
 
                     foreach (string ssfn in searchfile)
@@ -69,35 +74,43 @@ namespace SteamSSFN
                             File.Delete(textBox2.Text + "\\" + ssfn);
                         }
                     }
+
+                    //下载ssfn
                     HttpDownloadFile("https://ssfnbox.com" + ssfnURL, "temp");
+                    //移动至steam根目录
                     File.Move("temp", textBox2.Text + "/" + textBox1.Text);
+                    //启动steam
                     if (checkBox1.Checked == true)
                     {
                         System.Diagnostics.Process.Start(textBox2.Text, "-noreactlogin");
                     }
                     MessageBox.Show("SSFN写入成功");
                 }
-                else {
-                    MessageBox.Show("您确定您选择的Steam路径准确无误？","错误");
+                else
+                {
+                    MessageBox.Show("您确定您选择的Steam路径准确无误？", "错误");
                 }
             }
         }
 
+        //启动时自动获取steam路径
         private void Form1_Load(object sender, EventArgs e)
         {
             string steamPath = GetSteamPath();
             if (string.IsNullOrEmpty(steamPath))
             {
-                MessageBox.Show("Steam路径获取失败，请手动获取","错误");
+                MessageBox.Show("Steam路径获取失败，请手动获取", "错误");
             }
-            else {
+            else
+            {
                 textBox2.Text = steamPath;
             }
         }
 
+        //手动选择steam路径
         private void button3_Click(object sender, EventArgs e)
         {
-            string FolderPath=string.Empty;
+            string FolderPath = string.Empty;
             FolderBrowserDialog dialog = new FolderBrowserDialog();
             dialog.Description = "选择Steam目录";
             if (dialog.ShowDialog() == DialogResult.OK)
@@ -107,6 +120,7 @@ namespace SteamSSFN
             textBox2.Text = FolderPath;
         }
 
+        //运行steam
         private void button2_Click(object sender, EventArgs e)
         {
             if (File.Exists(textBox2.Text + "/steam.exe"))
@@ -119,14 +133,11 @@ namespace SteamSSFN
             }
         }
 
-        private void label4_Click(object sender, EventArgs e)
-        {
-
-        }
-
+        //一键上号
         private void button1_Click_1(object sender, EventArgs e)
         {
             if (File.Exists("temp")) File.Delete("temp");
+            //分割字符串
             string key = textBox3.Text;
             string[] keyarr = key.Split(new string[] { "----" }, StringSplitOptions.None);
             if (keyarr.Length >= 3)
@@ -134,6 +145,7 @@ namespace SteamSSFN
                 string login = keyarr[0].ToString();
                 string password = keyarr[1].ToString();
                 string loginssfn = keyarr[2].ToString();
+                //获取ssfn下载地址
                 HttpDownloadFile("https://ssfnbox.com/download/" + loginssfn, "temp");
                 string ssfnURL = GetBetweenStr(File.ReadAllLines("temp")[12], "window.location.assign(\"", "\"");
                 if (string.IsNullOrEmpty(ssfnURL))
@@ -142,8 +154,10 @@ namespace SteamSSFN
                 }
                 else
                 {
+                    //判断steam目录是否正确
                     if (File.Exists(textBox2.Text + "/steam.exe"))
                     {
+                        //在steam根目录下搜索并删除ssfn
                         string[] searchfile = Directory.GetFiles(textBox2.Text, "ssfn*").Select(path => Path.GetFileName(path)).ToArray();
 
                         foreach (string ssfn in searchfile)
@@ -153,22 +167,21 @@ namespace SteamSSFN
                                 File.Delete(textBox2.Text + "\\" + ssfn);
                             }
                         }
+
+                        //下载ssfn
                         HttpDownloadFile("https://ssfnbox.com" + ssfnURL, "temp");
+                        //移动至steam根目录
                         File.Move("temp", textBox2.Text + "/" + loginssfn);
+                        //启动steam
                         if (checkBox1.Checked == true)
                         {
-                            System.Diagnostics.Process.Start(textBox2.Text + "/steam.exe", "-login " + login +" "+ password+ " -noreactlogin -rememberpassword -windowed -bigpicture ");
+                            System.Diagnostics.Process.Start(textBox2.Text + "/steam.exe", "-login " + login + " " + password + " -noreactlogin -rememberpassword -windowed -bigpicture ");
                         }
                         MessageBox.Show("一键上号成功");
                     }
                     else MessageBox.Show("您确定您选择的Steam路径准确无误？", "错误");
-                }   
+                }
             }
-        }
-
-        private void checkBox1_CheckedChanged(object sender, EventArgs e)
-        {
-
         }
     }
 }
